==> code/1_LR_NB.ipynb <==
#Load the libraries
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import nltk
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from bs4 import BeautifulSoup
import spacy
import re,string,unicodedata
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from textblob import TextBlob
from textblob import Word
from imblearn.over_sampling import SMOTE
from collections import Counter
from sklearn.svm import LinearSVC
from sklearn.model_selection import train_test_split,cross_val_score, KFold

from sklearn.metrics import accuracy_score, f1_score, confusion_matrix,classification_report
from sklearn.metrics import recall_score,precision_score,precision_recall_fscore_support
from nltk.corpus import wordnet
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

from sklearn.pipeline import Pipeline
from sklearn.multiclass import OneVsRestClassifier

nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('averaged_perceptron_tagger')

import pandas as pd
import numpy as np
import nltk
from nltk.corpus import wordnet
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize
import re
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('averaged_perceptron_tagger')

from google.colab import drive
drive.mount('/content/gdrive', force_remount=True)

train_data = '/content/gdrive/MyDrive/NLP_Project/labels.csv'
test_data = '/content/gdrive/MyDrive/NLP_Project/2000_testdata.csv'



#basic preprocessing

#removing punctuation
def remove_punct_numbers(words):
    return [word.lower() for word in words if word.isalpha()]

# remove stop words
stop_words = nltk.corpus.stopwords.words('english')
def remove_stopwords(words):
    return [word for word in words if word not in stop_words]


# lemmatize
lematizer = WordNetLemmatizer()

def get_wordnet_pos(word):
    tag = nltk.pos_tag([word])[0][1][0].upper()
    tag_dict = {"J": wordnet.ADJ,
                "N": wordnet.NOUN,
                "V": wordnet.VERB,
                "R": wordnet.ADV}

    return tag_dict.get(tag, wordnet.NOUN)


def lematize(words):
    words = [lematizer.lemmatize(word,pos=get_wordnet_pos(word)) for word in words]
    #words = [lematizer.lemmatize(word,pos='v') for word in words]
    return words


#Preprocessing train data and test data 

def input_func(file, label):
    #input_path = input('/content/labels.csv')
    #input_path = input_path.replace ("/","//")
    df = pd.read_csv(file)
    df = df.dropna()
    df['processed'] = df[label].apply(lambda x:nltk.word_tokenize(x))
    df['processed'] = df['processed'].apply(remove_punct_numbers)
    df['processed'] = df['processed'].apply(remove_stopwords)
    df['processed'] = df['processed'].apply(lematize)
    return df

df = input_func(train_data, 'text_ocr')
df.to_csv('train_data_processed.csv')

df = input_func(test_data, 'OCR_extracted_text')
df.to_csv('test_data_processed.csv')

# TASK1 using LR/NB

#reading csv file
def read_corpus(file):
  return pd.read_csv(file)
  
train_data = '/content/gdrive/MyDrive/NLP_Project/train_data_processed.csv'
test_data = '/content/gdrive/MyDrive/NLP_Project/test_data_processed.csv'
test_true = '/content/gdrive/MyDrive/NLP_Project/Test_Actual_Final.csv'


#Dataset containing true values 

true_df = read_corpus(test_true)
#Extracting the first digit (1, 0 , -1) from Labels 

true_df['Sentiment'] = true_df['Labels'].str.split('_').str[0]
true_df['Sentiment'] = true_df['Sentiment'].astype(int)
true_df.head()

#Dataset containing the Train data
train_df = read_corpus(train_data)

train_df.head()

train_df['overall_sentiment'].value_counts()


#converting categorical to numerical 

train_df['overall_sentiment'].replace({'very_negative': -1, 'negative': -1, 'neutral': 0, 'positive': 1, 'very_positive': 1}, inplace=True)
train_df['overall_sentiment'].value_counts()

print(train_df.humour.value_counts())
print(train_df.sarcasm.value_counts())
print(train_df.offensive.value_counts())
print(train_df.motivational.value_counts())

Task B:

Not humorous => 0 and Humorous (funny, very funny, hilarious) => 1
Not Sarcastic => 0 and Sarcastic (general, twisted meaning, very twisted) => 1
Not offensive => 0 and Offensive (slight, very offensive, hateful offensive) => 1
Not Motivational => 0 and Motivational => 1


train_df.humour[train_df['humour']!='not_funny']= 1
train_df.humour[train_df['humour']=='not_funny']= 0
train_df.sarcasm[train_df['sarcasm']!='not_sarcastic']= 1
train_df.sarcasm[train_df['sarcasm']=='not_sarcastic']= 0
train_df.offensive[train_df['offensive']!='not_offensive']= 1
train_df.offensive[train_df['offensive']=='not_offensive']= 0
train_df.motivational[train_df['motivational']!='not_motivational']= 1
train_df.motivational[train_df['motivational']=='not_motivational']= 0

train_df.head()

print(train_df.humour.value_counts())
print(train_df.sarcasm.value_counts())
print(train_df.offensive.value_counts())
print(train_df.motivational.value_counts())

#Dataset containing the processed text of test data

test_df = read_corpus(test_data)
test_df.head()

#required data

#for training
train_text = train_df['processed']  #X_train
sentiment = train_df['overall_sentiment'] #Y_train

#for testing
test_text = test_df['processed'] #X_test

#for predicting
actual_test_sentiment = true_df['Sentiment'] #Y_test

#1. Count vectorizer for bag of words

cv = CountVectorizer(min_df=0,max_df=1,binary=False,ngram_range=(1,3))
cv_train = cv.fit_transform(train_text)
cv_test = cv.transform(test_text)



#2. using TFIDF

vectorizer = TfidfVectorizer(ngram_range=(1, 3))

cv_train = vectorizer.fit_transform(train_text)    
cv_test = vectorizer.transform(test_text)


#training the model using LR
model = LogisticRegression(penalty='l2', max_iter=500, C=1, random_state=42)

#model = MultinomialNB()

#Fitting the model for Bag of words
model.fit(cv_train, sentiment)

print("\n Training Accuracy: ", model.score(cv_train, sentiment)*100)


#making prediction
predicted_test_sentiment = model.predict(cv_test)

#calculating accuracy

print(' Test accuracy is {}'.format(accuracy_score(actual_test_sentiment, predicted_test_sentiment) * 100))
print(" F1 Score: {:.2f}".format(f1_score(actual_test_sentiment, predicted_test_sentiment, average='macro') * 100))
print(" Precision Score: {:.2f}".format(precision_score(actual_test_sentiment, predicted_test_sentiment, average='macro') * 100))
print(" Recall Score: {:.2f}".format(recall_score(actual_test_sentiment, predicted_test_sentiment, average='macro') * 100))

report = classification_report(actual_test_sentiment, predicted_test_sentiment)
print(report)

# TASK2 using LR/NB

## Import The Dataset  
train_df=pd.read_csv(train_data)
test_df,true_df=pd.read_csv(test_data),pd.read_csv(test_true)

train_df.humour[train_df['humour']!='not_funny']='1'
train_df.humour[train_df['humour']=='not_funny']='0'
train_df.sarcasm[train_df['sarcasm']!='not_sarcastic']='1'
train_df.sarcasm[train_df['sarcasm']=='not_sarcastic']='0'
train_df.offensive[train_df['offensive']!='not_offensive']='1'
train_df.offensive[train_df['offensive']=='not_offensive']='0'
train_df.motivational[train_df['motivational']!='not_motivational']='1'
train_df.motivational[train_df['motivational']=='not_motivational']='0'

def Multilabel(classifier,train_df,test_df,true_df):
    X_train=train_df['processed']
    X_test=test_df['processed']
    
    #### Create Bag of words 
    cv = CountVectorizer(min_df=0,max_df=1,binary=False,ngram_range=(1,3))
    cv_train = cv.fit_transform(X_train)
    cv_test = cv.transform(X_test)
    
    if classifier.lower()=='lr':
        # Using pipeline for applying logistic regression and one vs rest classifier
        model_pipeline = Pipeline([
                ('clf', OneVsRestClassifier(LogisticRegression(solver='sag'), n_jobs=-1)),
            ])
    elif classifier.lower()=='nb':
        # Using pipeline for applying Naive Bayes and one vs rest classifier
        model_pipeline  = Pipeline([
                ('clf', OneVsRestClassifier(MultinomialNB(), n_jobs=-1)),
            ])
        
    
    categories={'humour':0,'sarcasm':1,'offensive':2,'motivational':3}
    

    for category in categories.keys():
        print('*Processing {} category...*'.format(category))
    
        # Training model pipeline on train data
        model_pipeline.fit(cv_train, train_df[category])
        
        # Applying k-Fold Cross Validation
        accuracies = cross_val_score(model_pipeline, X = cv_train, y = train_df[category], cv=KFold(n_splits=5))
        print("\n Accuracy: {:.2f} %".format(accuracies.mean()*100))
        print(" Standard Deviation: {:.2f} %".format(accuracies.std()*100))
        
        # calculating test accuracy
        predicted = model_pipeline.predict(cv_test)
        y_test=true_df['Labels'].str.split('_').str[1].str[categories[category]]
        report = classification_report(y_test, predicted)
        
        print(report)
        print(' Test accuracy is {}'.format(accuracy_score(y_test,predicted)))
        print(" F1 Score: {:.2f}".format(f1_score(y_test, predicted, average='macro') * 100))
        print(" Precision Score: {:.2f}".format(precision_score(y_test, predicted, average='weighted') * 100))
        print(" Recall Score: {:.2f}".format(recall_score(y_test, predicted, average='weighted') * 100))
        print("\n")

Multilabel('LR',train_df,test_df,true_df)

Multilabel('NB',train_df,test_df,true_df)
