==> code/2_FFNN.ipynb <==
# Preparing the datasets for FFNN

#Load the libraries
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import nltk
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from bs4 import BeautifulSoup
import spacy
import re,string,unicodedata
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from textblob import TextBlob
from textblob import Word
from imblearn.over_sampling import SMOTE
from collections import Counter
from sklearn.svm import LinearSVC
from sklearn.model_selection import train_test_split

from sklearn.metrics import accuracy_score, f1_score, confusion_matrix,classification_report
from sklearn.metrics import recall_score,precision_score,precision_recall_fscore_support

from google.colab import drive
drive.mount('/content/gdrive', force_remount=True)

#reading csv file
def read_corpus(file):
  return pd.read_csv(file)
  
train_data = '/content/gdrive/MyDrive/NLP_Project/train_data_processed.csv'
test_data = '/content/gdrive/MyDrive/NLP_Project/test_data_processed.csv'
test_true = '/content/gdrive/MyDrive/NLP_Project/Test_Actual_Final.csv'

#Dataset containing the meme ground truth 

true_df = read_corpus(test_true)
#Extracting the first digit (1, 0 , -1) from Labels 

true_df['Sentiment'] = true_df['Labels'].str.split('_').str[0]
true_df['Sentiment'] = true_df['Sentiment'].astype(int)
true_df.head()

#Dataset containing the Train data
train_df = read_corpus(train_data)

train_df.head()

train_df['overall_sentiment'].value_counts()

#converting categorical to numerical 

train_df['overall_sentiment'].replace({'very_negative': -1, 'negative': -1, 'neutral': 0, 'positive': 1, 'very_positive': 1}, inplace=True)
train_df['overall_sentiment'].value_counts()

print(train_df.humour.value_counts())
print(train_df.sarcasm.value_counts())
print(train_df.offensive.value_counts())
print(train_df.motivational.value_counts())

train_df.humour[train_df['humour']!='not_funny']= 1
train_df.humour[train_df['humour']=='not_funny']= 0
train_df.sarcasm[train_df['sarcasm']!='not_sarcastic']= 1
train_df.sarcasm[train_df['sarcasm']=='not_sarcastic']= 0
train_df.offensive[train_df['offensive']!='not_offensive']= 1
train_df.offensive[train_df['offensive']=='not_offensive']= 0
train_df.motivational[train_df['motivational']!='not_motivational']= 1
train_df.motivational[train_df['motivational']=='not_motivational']= 0

train_df.head()

print(train_df.humour.value_counts())
print(train_df.sarcasm.value_counts())
print(train_df.offensive.value_counts())
print(train_df.motivational.value_counts())

#Dataset containing the processed text of test data

test_df = read_corpus(test_data)
test_df.head()

# Task1 using FFNN 

import torch.nn as nn
import torch.nn.functional as F
import torch.optim as optim
import torch
from gensim.utils import simple_preprocess
import gensim
from gensim.parsing.preprocessing import remove_stopwords, STOPWORDS
from gensim.parsing.porter import PorterStemmer
porter_stemmer = PorterStemmer()      


# Use cuda if present
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print("Device available for running: ")
print(device)

train_df_sub = train_df
train_df_sub.head()

#preprocessing train data
def preprocess(train_df_sub, label):
  # Tokenize the text column to get the new column 'tokenized_text'
  train_df_sub['tokenized_text'] = [simple_preprocess(line, deacc=True) for line in train_df_sub[label]] 

  #remove stop words
  train_df_sub['stop_tokens'] = [[remove_stopwords(word) for word in tokens] for tokens in train_df_sub['tokenized_text'] ]

  #remove punctuations
  train_df_sub['rem_punct_tokens'] = [[word.lower() for word in tokens if word.isalpha()] for tokens in train_df_sub['stop_tokens'] ]

  # Get the stemmed_tokens
  train_df_sub['pre_tokens'] = [[porter_stemmer.stem(word) for word in tokens] for tokens in train_df_sub['rem_punct_tokens'] ]
  
  return train_df_sub
  #train_df_sub['pre_tokens'].head(10)

train_df_sub = preprocess(train_df_sub, 'text_ocr')

train_df_sub.to_csv('train_data_processed.csv')

train_df_sub.head()

#The FFNN model

class FeedforwardNeuralNetModel(nn.Module):
    def __init__(self, input_size, hidden_size, num_classes):
        super(FeedforwardNeuralNetModel, self).__init__()
        
        # Linear function 1: 
        self.l1 = nn.Linear(input_size, hidden_size) 
        # Non-linearity 1
        self.relu1 = nn.ReLU()

        self.l2 = nn.Linear(hidden_size, hidden_size)
        # Non-linearity 2
        self.relu2 = nn.ReLU()

        # Linear function 2 
        self.l3 = nn.Linear(hidden_size, num_classes)  

    def forward(self, x):
        out = self.l1(x)
        out = self.relu1(out)
        out = self.l2(out)
        out = self.relu2(out)
        out = self.l3(out)
        return out

#Creating a dictionary of words 

from gensim import corpora
# Function to return the dictionary either with padding word or without padding
def make_dict(train_df_sub, padding=True):
    if padding:
        print("Dictionary with padded token added")
        review_dict = corpora.Dictionary([['pad']])
        review_dict.add_documents(train_df_sub['pre_tokens'])
    else:
        print("Dictionary without padding")
        review_dict = corpora.Dictionary(train_df_sub['pre_tokens'])
    return review_dict

# Make the dictionary without padding for the basic models
review_dict = make_dict(train_df_sub, padding=False)

#creating bag of words 

VOCAB_SIZE = len(review_dict)

# Function to make bow vector to be used as input to network
def make_bow_vector(review_dict, sentence):
    vec = torch.zeros(VOCAB_SIZE, dtype=torch.float64, device=device)
    for word in sentence:
        vec[review_dict.token2id[word]] += 1
    return vec.view(1, -1).float()

def make_bow_vector1(review_dict, sentence):
    vec = torch.zeros(VOCAB_SIZE, dtype=torch.float64, device=device)
    for word in sentence:
      if word in review_dict.keys():
        vec[review_dict.token2id[word]] += 1
    return vec.view(1, -1).float()

# Function to get the output tensor
def make_target(label):
    if label == -1:
        return torch.tensor([2], dtype=torch.long, device=device)
    elif label == 0:
        return torch.tensor([0], dtype=torch.long, device=device)
    else:
        return torch.tensor([1], dtype=torch.long, device=device)

#Training the FFNN model

VOCAB_SIZE = len(review_dict)

input_size = VOCAB_SIZE
hidden_size = 500
num_classes = 3
num_epochs = 100

ff_nn_bow_model = FeedforwardNeuralNetModel(input_size, hidden_size, num_classes)
ff_nn_bow_model.to(device)

loss_function = nn.CrossEntropyLoss()
optimizer = optim.SGD(ff_nn_bow_model.parameters(), lr=0.01)
#optimizer = torch.optim.Adam(ff_nn_bow_model.parameters(), lr=0.001)


X_train = train_df_sub[['pre_tokens']]
sentiment = train_df_sub['overall_sentiment']

# Start training
for epoch in range(num_epochs):
    if (epoch+1) % 25 == 0:
        print("Epoch completed: " + str(epoch+1))
    train_loss = 0
    for index, row in X_train.iterrows():
        # Clearing the accumulated gradients
        optimizer.zero_grad()

        # Make the bag of words vector for stemmed tokens 
        bow_vec = make_bow_vector(review_dict, row['pre_tokens'])
       
        # Forward pass to get output
        probs = ff_nn_bow_model(bow_vec)
        

        # Get the target label
        lab = train_df_sub['overall_sentiment'][index]
        target = make_target(lab)

        # Calculate Loss: softmax --> cross entropy loss
        loss = loss_function(probs, target)
        # Accumulating the loss over time
        train_loss += loss.item()

        # Getting gradients w.r.t. parameters
        loss.backward()

        # Updating parameters
        optimizer.step()

test_df_sub = test_df
#test_df_sub

#Preprocessing of test data
test_df_sub = preprocess(test_df_sub, 'OCR_extracted_text')
test_df_sub.to_csv('test_data_processed.csv')

test_df_sub.head()

from sklearn.metrics import classification_report
bow_ff_nn_predictions = []
original_lables_ff_bow = []

with torch.no_grad():
    for index, row in test_df_sub.iterrows():
        bow_vec = make_bow_vector1(review_dict, row['pre_tokens'])
        probs = ff_nn_bow_model(bow_vec)
        bow_ff_nn_predictions.append(torch.argmax(probs, dim=1).cpu().numpy()[0])
        original_lables_ff_bow.append(make_target(true_df['Sentiment'][index]).cpu().numpy()[0])
print(classification_report(original_lables_ff_bow,bow_ff_nn_predictions))
print(' Test accuracy is {}'.format(accuracy_score(original_lables_ff_bow,bow_ff_nn_predictions) * 100))
print(" F1 Score: {:.2f}".format(f1_score(original_lables_ff_bow,bow_ff_nn_predictions, average='macro') * 100))
print(" Precision Score: {:.2f}".format(precision_score(original_lables_ff_bow,bow_ff_nn_predictions, average='macro') * 100))
print(" Recall Score: {:.2f}".format(recall_score(original_lables_ff_bow,bow_ff_nn_predictions, average='macro') * 100))
print("\n")

# Task2 using FFNN

train_df_sub.head()

# Function to get the output tensor
def make_target_2(label):
    if label == 1:
        return torch.tensor([1], dtype=torch.long, device=device)
    elif label == 0:
        return torch.tensor([0], dtype=torch.long, device=device)

X_train = train_df_sub[['pre_tokens']]

categories={'humour':0,'sarcasm':1,'offensive':2,'motivational':3}

for category in categories.keys():
  print('******Processing {} category...******'.format(category))

  # Start training
  for epoch in range(num_epochs):
      #if (epoch+1) % 25 == 0:
      #   print("Epoch completed: " + str(epoch+1))
      train_loss = 0
      for index, row in X_train.iterrows():
          # Clearing the accumulated gradients
          optimizer.zero_grad()

          # Make the bag of words vector for stemmed tokens 
          bow_vec = make_bow_vector(review_dict, row['pre_tokens'])
        
          # Forward pass to get output
          probs = ff_nn_bow_model(bow_vec)
          
          # Get the target label
          lab = train_df_sub[category][index]
          target = make_target_2(lab)
          #print(lab)
          #print(target)

          # Calculate Loss: softmax --> cross entropy loss
          loss = loss_function(probs, target)
          # Accumulating the loss over time
          train_loss += loss.item()

          # Getting gradients w.r.t. parameters
          loss.backward()

          # Updating parameters
          optimizer.step()

  bow_ff_nn_predictions = []
  original_lables_ff_bow = []

  with torch.no_grad():
      for index, row in test_df_sub.iterrows():
          bow_vec = make_bow_vector1(review_dict, row['pre_tokens'])
          probs = ff_nn_bow_model(bow_vec)
          bow_ff_nn_predictions.append(torch.argmax(probs, dim=1).cpu().numpy()[0])
          lab = true_df['Labels'].str.split('_').str[1].str[categories[category]][index]
          lab = int(lab)
          #print(lab)
          original_lables_ff_bow.append(make_target_2(lab).cpu().numpy()[0])
  print(classification_report(original_lables_ff_bow,bow_ff_nn_predictions))
  print(' Test accuracy is {}'.format(accuracy_score(original_lables_ff_bow,bow_ff_nn_predictions) * 100))
  print(" F1 Score: {:.2f}".format(f1_score(original_lables_ff_bow,bow_ff_nn_predictions, average='macro') * 100))
  print(" Precision Score: {:.2f}".format(precision_score(original_lables_ff_bow,bow_ff_nn_predictions, average='macro') * 100))
  print(" Recall Score: {:.2f}".format(recall_score(original_lables_ff_bow,bow_ff_nn_predictions, average='macro') * 100))
  print("\n")
