==> code/4_Sampling.ipynb <==
#Importing all the required libraries
import pandas as pd
import numpy as np
import nltk
from nltk.corpus import wordnet
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB

from sklearn.metrics import classification_report, confusion_matrix, accuracy_score

import torch
import torch.nn as nn
import torch.nn.functional as F
import torch.optim as optim
from torch.utils.data import (TensorDataset, DataLoader, RandomSampler, SequentialSampler)

from imblearn.over_sampling import SMOTE, RandomOverSampler
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import train_test_split
from sklearn.metrics import precision_score, recall_score, f1_score

import warnings
warnings.filterwarnings("ignore")

#Downloading NLTK libraries
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('averaged_perceptron_tagger')

# Use cuda if present
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print("Device available for running: ")
print(device)

#Root folder for data files (for storage and retrieval)
GDRIVE_PROJECT_FOLDER = '/content/gdrive/MyDrive/NLP_Project/'

#Mount the google drive to access data files
from google.colab import drive
drive.mount('/content/gdrive', force_remount=True)

#Path variables for the train and test data files
train_data = GDRIVE_PROJECT_FOLDER+'train_data_processed.csv'
test_data = GDRIVE_PROJECT_FOLDER+'test_data_processed.csv'
test_true = GDRIVE_PROJECT_FOLDER+'Test_Actual_Final.csv'

#Dataset containing the meme ground truth 
true_df = pd.read_csv(test_true)
true_df.head()

#Extracting the first digit (1, 0 , -1) from Labels 
true_df['Sentiment'] = true_df['Labels'].str.split('_').str[0]
true_df['Sentiment'] = true_df['Sentiment'].astype(int)
true_df.head()

#Dataset containing the Train data
train_df = pd.read_csv(train_data, converters={'pre_tokens': eval, 'processed': eval})
train_df.head()

#Checking the class labels balance in training dataset for task 1 (identifying meme as positive/negative/neutral - (1/-1/0))
train_df['overall_sentiment'].value_counts()

#Dataset containing the processed text of test data
test_df = pd.read_csv(test_data, converters={'pre_tokens': eval, 'processed': eval})
test_df.head()

#Creating a duplicate copy of train dataframe, so that modifications can be done in copy df if needed
train_df_sub = train_df
train_df_sub.head()

# Handling class imbalance using Random Oversampling and SMOTE techniques
# with Logistic Regression model

'''dummy_param takes a word as parameter and returns the same word
   This function is useful when the sentences are already tokenized
   and these tokenized sentences are given as input to 
   CountVectorizer() and TfidfVectorizer() for creating BOW and TF-IDF'''
def dummy_param(doc):
    return doc

'''since the labels and sentences of test set are in different files (dataframes)
   both these dataframes are merged to form test dataset'''
test_df_sub = pd.merge(test_df[['Unnamed: 0','processed']], true_df[['Unnamed: 0','Sentiment']], on='Unnamed: 0')
test_df_sub.rename(columns = {"Sentiment": "sentiment"}, inplace=True)

'''converting text to word count vectors - BOW
   tokenizer and preprocessor - are set to callable function which takes a word and return the same word
                                since the sentences are already tokenized, parameters  
   ngrame_range - (1,3) considers unigram, bigram and trigram'''
cv = CountVectorizer(min_df=0,max_df=1,binary=False,ngram_range=(1,3), tokenizer = dummy_param, analyzer="word", preprocessor=dummy_param)
cv_train = cv.fit_transform(train_df_sub['processed'])
cv_test = cv.transform(test_df_sub['processed'])

'''Convert a collection of raw documents to a matrix of TF-IDF features
   tokenizer and preprocessor - are set to callable function which takes a word and return the same word
                                since the sentences are already tokenized, parameters  
   ngrame_range - (1,3) considers unigram, bigram and trigram'''
training_text = pd.Series([sent for sent in train_df_sub['pre_tokens']])
training_target = pd.Series([label for label in train_df_sub['overall_sentiment']])
tv = TfidfVectorizer(stop_words=None, max_features=100000, ngram_range=(1, 3), analyzer='word', tokenizer=dummy_param, preprocessor=dummy_param)
training_tfidf = tv.fit_transform(training_text)

#preparing testing dataset from the test dataframe
testing_text = pd.Series([sent for sent in test_df_sub['processed']])
testing_target = pd.Series([label for label in test_df_sub['sentiment']])
testing_tfidf = tv.transform(testing_text)

#creating an object for RandomOverSampler for random oversampling
ros = RandomOverSampler(random_state=777)

#creating an object for SMOTE for SMOTE oversampling
smt = SMOTE(random_state=777, k_neighbors=5)


samplings = {'No Sampling':'no_sampling', 'ROS':ros, 'SMOTE':smt}
representations = {'TFIDF':training_tfidf, 'BOW':cv_train}

#this loop runs for 3 times for each of - No Sampling, ROS, SMOTE
for sampling_key in samplings:
  sampling = samplings[sampling_key]

  #this loop runs twice for each of the above sampling technique with tfidf features and bow separately
  for representation_key in representations:
    representation = representations[representation_key]

    #stratified k fold is used to have balanced number of labels in each of the folds
    kfold = StratifiedKFold(n_splits=5, shuffle=True, random_state=777)
    accuracy = []
    precision = []
    recall = []
    f1 = []

    X = representation
    Y = training_target

    average_method = 'macro'
    model = LogisticRegression(penalty='l2', max_iter=500, C=1, random_state=42)

    #for each of k folds, model is trained with train fold and tested with test fold
    #precision, recall, f1 and accuracy are performance measures
    for train, test in kfold.split(X, Y):

        #to decide whether to perform oversampling for train dataset or not
        if sampling =='no_sampling':
          X_train = representation
          y_train = training_target
        else:
          X_train, y_train = sampling.fit_sample(X[train], Y[train])
          
        model.fit(X_train, y_train)
        prediction = model.predict(X[test])
        scores = model.score(X[test],Y[test])
        
        accuracy.append(scores * 100)
        precision.append(precision_score(Y[test], prediction, average=average_method)*100)
        recall.append(recall_score(Y[test], prediction, average=average_method)*100)
        f1.append(f1_score(Y[test], prediction, average=average_method)*100)

    print("Representation: {} | Sampling: {}".format(representation_key, sampling_key))
    print("\nTrain dataset")
    print("accuracy: %.2f%% (+/- %.2f%%)" % (np.mean(accuracy), np.std(accuracy)))
    print("precision: %.2f%% (+/- %.2f%%)" % (np.mean(precision), np.std(precision)))
    print("recall: %.2f%% (+/- %.2f%%)" % (np.mean(recall), np.std(recall)))
    print("f1 score: %.2f%% (+/- %.2f%%)" % (np.mean(f1), np.std(f1)))

    #to select the feature extraction method for test set
    print("\nTest dataset")
    if representation_key == 'TFIDF':
      X_test = testing_tfidf
    else:
      X_test = cv_test
    y_test = testing_target

    prediction = model.predict(X_test)
    scores = model.score(X_test,y_test)
    print("Accuracy: %.2f%%" % (scores * 100))
    print("precision: %.2f%%" % (precision_score(y_test, prediction, average=average_method)*100))
    print("recall: %.2f%%" % (recall_score(y_test, prediction, average=average_method)*100))
    print("f1 score: %.2f%%" % (f1_score(y_test, prediction, average=average_method)*100))
    print("-"*100)
