==> code/5_BERT.ipynb <==
# Preparing the datasets for BERT

#Load the libraries
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import nltk
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from bs4 import BeautifulSoup
import spacy
import re,string,unicodedata
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from textblob import TextBlob
from textblob import Word
from imblearn.over_sampling import SMOTE
from collections import Counter
from sklearn.svm import LinearSVC
from sklearn.model_selection import train_test_split

from sklearn.metrics import accuracy_score, f1_score, confusion_matrix,classification_report
from sklearn.metrics import recall_score,precision_score,precision_recall_fscore_support

from google.colab import drive
drive.mount('/content/gdrive', force_remount=True)

#reading csv file
def read_corpus(file):
  return pd.read_csv(file)
  
train_data = '/content/gdrive/MyDrive/NLP_Project/train_data_processed.csv'
test_data = '/content/gdrive/MyDrive/NLP_Project/test_data_processed.csv'
test_true = '/content/gdrive/MyDrive/NLP_Project/Test_Actual_Final.csv'

#Dataset containing the meme ground truth 

true_df = read_corpus(test_true)
#Extracting the first digit (1, 0 , -1) from Labels 

true_df['Sentiment'] = true_df['Labels'].str.split('_').str[0]
true_df['Sentiment'] = true_df['Sentiment'].astype(int)
true_df.head()

#Dataset containing the Train data
train_df = read_corpus(train_data)

train_df.head()

train_df['overall_sentiment'].value_counts()

#converting categorical to numerical 

train_df['overall_sentiment'].replace({'very_negative': -1, 'negative': -1, 'neutral': 0, 'positive': 1, 'very_positive': 1}, inplace=True)
train_df['overall_sentiment'].value_counts()

print(train_df.humour.value_counts())
print(train_df.sarcasm.value_counts())
print(train_df.offensive.value_counts())
print(train_df.motivational.value_counts())

train_df.humour[train_df['humour']!='not_funny']= 1
train_df.humour[train_df['humour']=='not_funny']= 0
train_df.sarcasm[train_df['sarcasm']!='not_sarcastic']= 1
train_df.sarcasm[train_df['sarcasm']=='not_sarcastic']= 0
train_df.offensive[train_df['offensive']!='not_offensive']= 1
train_df.offensive[train_df['offensive']=='not_offensive']= 0
train_df.motivational[train_df['motivational']!='not_motivational']= 1
train_df.motivational[train_df['motivational']=='not_motivational']= 0

train_df.head()

print(train_df.humour.value_counts())
print(train_df.sarcasm.value_counts())
print(train_df.offensive.value_counts())
print(train_df.motivational.value_counts())

#Dataset containing the processed text of test data

test_df = read_corpus(test_data)
test_df.head()

# Task1 using BERT

pip install transformers

from transformers import BertTokenizer, TFBertForSequenceClassification
from transformers import InputExample, InputFeatures
import tensorflow as tf
import pandas as pd
import numpy as np

train_df_bert = train_df
train_df_bert.head()

train_df_bert['overall_sentiment'].value_counts()

train_df_bert['overall_sentiment'].replace({-1: 2}, inplace=True)
train_df_bert['overall_sentiment'].value_counts()

test_df_bert = test_df
test_df_bert.head()

true_df_bert = true_df
true_df_bert.head()

text = train_df_bert['text_ocr'].values.tolist()
labels = train_df_bert['overall_sentiment'].tolist()

from sklearn.model_selection import train_test_split
training_sentences, validation_sentences, training_labels, validation_labels = train_test_split(text, labels, test_size = 0.1)

#Assign tokenizer object to the tokenizer class

tokenizer = BertTokenizer.from_pretrained("bert-base-uncased")
#tokenizer = DistilBertTokenizerFast.from_pretrained('distilbert-base-uncased')

tokenizer([training_sentences[0]], truncation=True,

                            padding=True, max_length=128)

#creating the encodings 

train_encodings = tokenizer(training_sentences,
                            truncation=True,
                            padding=True)
val_encodings = tokenizer(validation_sentences,
                            truncation=True,
                            padding=True)
train_dataset = tf.data.Dataset.from_tensor_slices((
                            dict(train_encodings),
                            training_labels
                            ))
val_dataset = tf.data.Dataset.from_tensor_slices((
                            dict(val_encodings),
                            validation_labels
                            ))

model = TFBertForSequenceClassification.from_pretrained("bert-base-uncased", num_labels=3)

#model = TFDistilBertForSequenceClassification.from_pretrained('distilbert-base-uncased',num_labels=3)

#training

optimizer = tf.keras.optimizers.Adam(learning_rate=5e-5, epsilon=1e-08)
model.compile(optimizer=optimizer, loss=model.compute_loss, metrics=['accuracy'])
model.fit(train_dataset.shuffle(100).batch(16),
          epochs=10,
          batch_size=16,
          validation_data=val_dataset.shuffle(100).batch(16))

model.save_pretrained("./sentiment")
loaded_model = TFBertForSequenceClassification.from_pretrained("./sentiment")

test_text = test_df_bert['OCR_extracted_text'].values.tolist()
test_text


predictions = []
labels = [0, 1, 2]
for test_sentence in test_text:
  predict_input = tokenizer.encode(test_sentence,
                                  truncation=True,
                                  padding=True,
                                  return_tensors="tf")

  tf_output = loaded_model.predict(predict_input)[0]
  tf_prediction = tf.nn.softmax(tf_output, axis=1)
  label = tf.argmax(tf_prediction, axis=1)
  #label = label.numpy()
  predictions.append(labels[label[0]])
  #print(labels[label[0]])


actual = true_df_bert['Sentiment'].values.tolist()

print(classification_report(actual, predictions))
print(' Test accuracy is {}'.format(accuracy_score(actual, predictions) * 100))
print(" F1 Score: {:.2f}".format(f1_score(actual, predictions, average='macro') * 100))
print(" Precision Score: {:.2f}".format(precision_score(actual, predictions, average='macro') * 100))
print(" Recall Score: {:.2f}".format(recall_score(actual, predictions, average='macro') * 100))
print("\n")

# Task2 using BERT 

tokenizer_2 = BertTokenizer.from_pretrained("bert-base-uncased")

model_2 = TFBertForSequenceClassification.from_pretrained("bert-base-uncased", num_labels=2)

#training 

text = train_df_bert['text_ocr'].values.tolist()
test_text = test_df_bert['OCR_extracted_text'].values.tolist() 

optimizer = tf.keras.optimizers.Adam(learning_rate=5e-5, epsilon=1e-08)


categories={'humour':0,'sarcasm':1,'offensive':2,'motivational':3}

for category in categories.keys():
  print('******Processing {} category...******'.format(category))

  labels = train_df_bert[category].tolist()
  training_sentences, validation_sentences, training_labels, validation_labels = train_test_split(text, labels, test_size = 0.1)


  train_encodings = tokenizer_2(training_sentences,
                              truncation=True,
                              padding=True)
  val_encodings = tokenizer_2(validation_sentences,
                              truncation=True,
                              padding=True)
  train_dataset = tf.data.Dataset.from_tensor_slices((
                              dict(train_encodings),
                              training_labels
                              ))
  val_dataset = tf.data.Dataset.from_tensor_slices((
                              dict(val_encodings),
                              validation_labels
                              ))

  model_2.compile(optimizer=optimizer, loss=model_2.compute_loss, metrics=['accuracy'])
  model_2.fit(train_dataset.shuffle(100).batch(16),
            epochs=3,
            batch_size=16,
            validation_data=val_dataset.shuffle(100).batch(16))
  

  #testing
  predictions = []
  actual = []
  labels = [0, 1]
  for test_sentence in test_text:
    predict_input = tokenizer_2.encode(test_sentence,
                                    truncation=True,
                                    padding=True,
                                    return_tensors="tf")

    tf_output = model_2.predict(predict_input)[0]
    tf_prediction = tf.nn.softmax(tf_output, axis=1)
    label = tf.argmax(tf_prediction, axis=1)
    predictions.append(labels[label[0]])

  for index, row in test_df_bert.iterrows():
    lab = true_df['Labels'].str.split('_').str[1].str[categories[category]][index]
    lab = int(lab)
    actual.append(lab)
    
  #actual = true_df_bert['Sentiment'].values.tolist()


  print(classification_report(actual, predictions))
  print(' Test accuracy is {}'.format(accuracy_score(actual, predictions) * 100))
  print(" F1 Score: {:.2f}".format(f1_score(actual, predictions, average='macro') * 100))
  print(" Precision Score: {:.2f}".format(precision_score(actual, predictions, average='macro') * 100))
  print(" Recall Score: {:.2f}".format(recall_score(actual, predictions, average='macro') * 100))
  print("\n")
